ApiChatbot/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain-core>=0.0.10
langgraph>=0.0.26
pydantic-ai>=0.0.1
nest-asyncio>=1.5.6
dateparser>=1.1.8
typing-extensions>=4.5.0
IPython>=8.0.0
aiohttp>=3.8.0
httpx>=0.24.0